requirements.txt
flask[async] == 3.1.0
requests == 2.32.3
python-dotenv ==1.0.1
aiohttp == 3.11.9
solders == 0.21.0
solana  == 0.35.1
pyjwt == 2.10.1
redis == 2.10.6